+++ app/layout.tsx
@@ -1,5 +1,5 @@
 import '@/app/ui/global.css';
-import { inter } from '@/app/ui/font';
+import { inter } from '@/app/ui/fonts';
  
 export default function RootLayout({
   children,
